Added a new property srio-ccsr-subwindow to support SRIO under HV.
On P4080, there is a single LIODN per both maintenance and other
general SRIO transactions. Maintenance transactions are translated
remotely into accesses to the CCSR space that must be authorized by
PAMU. There is not enough flexibility in defining a subwindow that
covers only the SRIO register space, so a larger area from the CCSR
space will be covered. However, the guest does not own this area.
For this purpose a special window is defined that must
include the SRIO CCSR registers:
		ccsr_space: ccsr_space {
			compatible = "dma-subwindow";
			guest-addr = <0xf 0xf0000000>;
			size = <0x0 0x10000000>;
			srio-ccsr-subwindow;
		};
This window translates to an entry in PAMU with a 1:1 mapping
between the guest address and the physical address.

On P5020, P3041 there are special LIODNs used only for maintenance
transactions and such a special subwindow is no longer needed.

Signed-off-by: Diana CRACIUN <stfd002@zro03-ws339.ea.freescale.net>
Signed-off-by: Laurentiu TUDOR <Laurentiu.Tudor@freescale.com>