Switched to initramfs from initrd
Modified simics script to load initramfs image.

Modified linux-hello and linux-partman to create initramfs images (initrd is still supported).

Modified partman to calculate and update the /chosen/linux,initrd nodes
according to the address and size of the initramfs image provided by load command.

A new flag (-r) specifies that the loaded image is a filesystem.

Signed-off-by: Adrian Raileanu <adrian.raileanu@freescale.com>