partman: update to new ioctl interface
The ioctl interface in the Linux kernel has changed slightly different, so
partman needs to use the new header file and make ioctl calls correctly.  In
particular, the union of all structures is gone, and the ioctl type has been
changed to 0xAF, which is does not conflict with other ioctls.

Signed-off-by: Timur Tabi <timur@freescale.com>