Fix dma test timeout happening on p5020 chips
Both partitions involved in the test trigger a pamu exception.
The first exception starts to get handled in the hv. The handler
prints a log message (an operation which takes _A_LOT_ of time) and
then clears the pamu error bits. However, the second exception was still
pending and is lost.

Add an arbitrary delay to first partition in order to
desynchronize the moment the error interrupts are triggered.

Signed-off-by: Laurentiu TUDOR <Laurentiu.Tudor@freescale.com>