Workaround for PAMU erratum A-003638
When a peripheral interface makes an invalid memory access,
an access violation interrupt is triggered. The interrupt
handler clears PICS[AVICS] to clear the interrupt and also
clears the appropriate bit from the PAACE entry to stop further
attempts from the peripheral interface that may cause further
access violations. Due to errata, clearing the valid bit in PAACE
entry will not stop further access violations from the interface.
Furthermore, if access violations continue to occur, the PAMU can
end up into a state where PICS[AVICS] is 0, but the access violation
interrupt is still fired.

Workaround:
When the access violation occurs,the access violation reporting is
disabled for this PAMU and it is the responsability of the guest to
re-enable DMA when the device is in a valid state (when no more illegal
memory accesses are performed). When the guest re-enables DMA, if that
liodn caused the access violation interrupts to be disabled, they will
be re-enabled.

Signed-off-by: Diana CRACIUN <Diana.Craciun@freescale.com>
Acked-by: Stuart Yoder <stuart.yoder@freescale.com>