Fixed error manger test due to workaround for PAMU A-003638 erratum
Due to erratum workaround, when an access violation is received, the
access violations for that PAMU are turned off until the guest re-enables
DMA for the faulty device.

For the error-manager test to work increased the delay between the two
slave guests such that the access violations are re-enabled when the
second guest performs the violation.

Signed-off-by: Diana CRACIUN <Diana.Craciun@freescale.com>