Fixed stash and snoop ID settings for Qman portals.
Qman portal has an cpu-handle property indicating the CPU
handle the portal is associated with. In the hypervisor
config tree there is a possibility to change this association
by specifying the vcpu property to the portal node.
If in the hardware device tree the portal is associated
with a cpu and in the hypervisor config tree with another cpu,
the stashing is not configured correctly because the Hypervisor
is using the association between portal and CPU from the HW config
tree instead of the one from the HV config tree.

Signed-off-by: Diana CRACIUN <Diana.Craciun@freescale.com>