Refactor number parsing utility functions
Removed error printing from the functions because there
are places where the caller does its own error printing
leading to duplicate error messages in the console.
Exported a function that can be used by the callers to
print number parsing errors.
Added error printing in a few places where the caller
either relied on the prints in the parsing functions or
error handling was missing altogether.

Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>