Fix issue in global handle allocation
The global handle allocation is maintained within a bit-array.
A set bit means that the handle equal with the index is in use
and a zero bit means that the handle is free.
The function that allocates the handles iterates this array by
using two imbricated 'for' loops: the outer loop iterates the
array elements and the inner loop the individual bits of an
element. When it finds a  free spot (a zero bit) it changes
it to 'in-use' (sets the bit) and breaks out of the inner loop.
The issue was that the function didn't also break from the outer
loop, causing several effects such as:
 - more handles marked as used
 - handles were allocated only from the last element of the array

Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>