Used the aliased MSIIR register to set up the MSI subwindow.
Topaz used the MSIIR address from the MSI bank to setup the
MSI window. This was a securiry hole as the MSI window pointed
to 0x4_1000 offset in MPIC memory map allowing the device to DMA
anything on that page.

The MSIIR register for each MSI bank is aliased to a different
address, each on their own page. With this patch Topaz is using
the address of this aliased address for setting up the MSI subwindow.

Signed-off-by: Diana CRACIUN <Diana.Craciun@freescale.com>