Added an LRAT unit test
The LRAT creates a number of PMAs (higher than the available
entries into the hardware LRAT) and performs some memory accesses.

It can be tested that the LRAT entries were correctly written
(otherwise an LRAT miss will be issued), but it is harder to test
that the LRAT is populated correctly/if the entries are correctly
evicted. However, this can be manually tested under simulator by
inspecting the entries after each access.

Signed-off-by: Diana CRACIUN <Diana.Craciun@freescale.com>