Added support for LRAT interrupt.
When the LRAT interrupt is received a LRAT entry is written
by the hypervisor to the hardware LRAT.

If the guest is trying to write to an invalid address (it does
not have permissions to write to the specified logical address),
no entry is added in the LRAT. In this case, the hypervisor is
writing the TLB entry on behalf of the guest and sets the VF bit.

Currently, the interrupt handler handles only misses generated
by the tlbwe instructions (no support for page table walk).

Signed-off-by: Diana CRACIUN <Diana.Craciun@freescale.com>