e6500: Don't touch L2 because it is no longer core local
Added a check at init time that sets a flag if the
L2 cache is core local if running on e500mc or e5500.
The cache functions check this flag, and if set also
act on the L2 cache. Otherwise, as is the case for e6500,
the L2 cache SPRs are left untouched.
This is needed because if we access the L2 cache SPRs on e6500
the hypervisor crashes with a program exception.

Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>