Disable TLB0 software cache for e6500
TLB0 cache is useful when the guest is not using hardware
page table walk or guest tlb management instructions. On
e6500 this should be an unlikely scenario. Adding support
for e6500 would imply some changes that may affect the
performance of the TLB0 cache on other platforms.
However, we may add support for TLB0 cache in the future if the
performance measurements do not indicate important degradation
or if there is a specific need for supporting the above
mentioned scenario.

Also enabled HES support as TLB0 cache scenario was the
only one that did not support HES. Any other scenario
do not need any particular changes in Topaz to support HES.

Signed-off-by: Diana CRACIUN <Diana.Craciun@freescale.com>