Added support for mux_server in unit tests
mux_server is started for each unit test run with
make test-run-test_name. The mux_server connects to
tio_console application from the T4 simulator and xterm
consoles are started for each partition and hypervisor
shell.

Also updated xtel.sh to allow running on local host, but
specifying a -exec command which is the case for T4
simulator.

Signed-off-by: Diana CRACIUN <Diana.Craciun@freescale.com>