test/nmi: remove hardcoded number of cpus
The test was expecting 5 CPUs. Removed this limitation by
handling dinamically the number of CPUs so that the test
can be run on chips with lesser number of cores.
Updated the number of CPUs allocated to part1 from 5 to 4
so that the test can run on platforms with 4 cores.
Updated test plan for 2041 and 3041 SoCs to include this test.

Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>