Fix TB based timeout loops on 64-bit
In 64-bit mode, reading the the TBL returns the entire
64-bit timebase value, not only the 32-bit low part.
In some timeout loops, the TBL was truncated by storing
it in an uint32_t.
This caused timeout loops coded as:
	uint32_t start_tb = mfspr(SPR_TBL);
	while (mfspr(SPR_TBL) - start_tb < timeout) {
		...
	}
to end prematurely and trigger false timeouts for
timebase values larger than 32 bits.
One visible result was L2 cache flush timeouts, indicated
by messages in the hv console:
	flush_disable_l2_cache: L2 cache flush timeout
	idle_loop: couldn't nap, using wait instead
The fix consists in storing the TBL in a register_t.

Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>