T4: Emulate several new e6500 instructions and SPRs
Most of the patch consists in simple emulation
of the new thread management instructions
added to the e6500 core plus returning LRATCFG = 0
when read by guests.
This is needed because there are guests that
detect e6500 cores by SVR and make the simplifying
assumption that the instrs and sprs are available.
For example, the linux kernel, in the implementation
available at the time this patch was written even
assumes 2 hw threads.

Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>