T4: Update the CCM driver to support T4240
The hardware platform allows the possibility to configure
coherency subdomains to limit the snooping traffic only
to the entities that are in the same coherency subdomain.
Before T4240 cores could be added in a coherency subdomain,
on T4240 clusters can be grouped in a coherency subdomain.
Also the SIDMR and CIDMR registers have a slightly changed
format.

Signed-off-by: Diana Craciun <Diana.Craciun@freescale.com>