Fix a (potential) race in dma init
gcpus array is initialized by each core
in its initialization code but when the
primary core initializes dma it assumes
that all the other cores are up and did
this initialization.
This was noticed when investigating
another issue involving secondary
threads not starting up and causing
the crash that this patch fixes.

Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>