Added .notes section to hypervisor lds section
The latest compiler adds by default the section .note.gnu.build-id
which contains the build ID which is a unique identification for the
build files.
The section was placed at address 0 which is the address where u-boot
expects to be the entry point into the Hypervisor.
A new section .notes was created which is explicitly placed at a
safe location.

Signed-off-by: Diana Craciun <Diana.Craciun@freescale.com>