e6500rev2: Emulate power management control register 0
Since the SPR is shared between threads
and threads are exposed as cores, on writes,
use the maximum values for the count periods
set by the threads. On reads, provide
the value that was originally written by
the thread (not the one with the maximum
counters).
The timeouts are enabled only when both threads
enable them.

Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>