Make the 'devices' container node mimic a 'soc' node
Add a couple of properties to the guest device
tree's "devices" node to make it look like a
"soc" node:
 - "device_type" set to "soc"
 - "bus-frequency" copied from the hw device tree
This is useful to the kernel when it tries to
retrieve the "system clock" by querying the
"bus-frequency" property from the "soc" node.
The value is used in various kernel drivers,
such as spi and watchdog.

Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>