Add checks for invalidated gtlb1 entries in several places
When running with "direct guest tlb mgmt" ON
an invalidate (tlbilx running natively) might
have removed a TLB1 entry from the hardware tlb1.
Add checks in tlb1 emulation code that iterates
through all guest TLB1 entries and, if the
corresponding hw entry was invalidated then
invalidate in the internal structs too.
This check was already being implemented in
update_dgtmi(), so extract it from there and
place it in a new function so it's callable
from the spots it's needed.

NOTE: This check isn't needed when the FAST_TLB1
Kconfig option is set because the TLB1 entries
handled on the fast path (non-IPROT ones) are kept
only in the hardware tlb1 while the others
(having IPROT set) are obviously not affected
by invalidates.

Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>