msi: Differentiate between mpic 4.3 and older ones
The new MSI block in MPIC 4.3 added the MSIIR1 register,
with a different layout, in order to support 16 MSIR
registers. The msi binding was also updated so that
the "reg" reflects the newly introduced MSIIR1 register.
This must be advertised to guests so they are
aware they must use the new MSIIR1 register layout so,
in this case, make the guest msi nodes compatible
string be "fsl,vmpic-msi-v4.3" instead of "fsl,vmpic-msi".

Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>