fix secondary gcpu spin-up for >=4GB guests
The spin table is hardcoded at 0xfffff000 so, for >=4GB guests
we end up with the impossible situation of having a guest physical
address (that is spin table's addr @0xfffff000) pointing to 2 real
physical addresses (the spin table memory and the guest pma).
In the case of >=4GB guests, fix by placing the spin table at the
end of the guest's addr space while for <4GB guests let the address
as before.

Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>