Ensure L1 cache is properly enabled after cache flush
In a multi-threaded system, it is possible for one thread of a CPU to
read the shared L1 data and instruction caches as disabled, leading
to the L1 cache getting permanently disabled for that core. Explicitly
re-enable both of the L1 caches if they were enabled before the cache
flush started.

Signed-off-by: Joe Schultz <jschultz@xes-inc.com>